+++ MasterEvent/UI/GmWindow.cs
@@ -27,7 +27,7 @@ public sealed partial class GmWindow : MasterEventWindowBase
     // Météo et temps
     private byte selectedWeatherId;
     private Dictionary<byte, string>? cachedWeatherList;
-    private ushort cachedTerritoryId;
+    private uint cachedTerritoryId;
     private int selectedHour = -1; // -1 = pas encore initialisé
 
     private static readonly HttpClient httpClient = new() { Timeout = TimeSpan.FromSeconds(5) };
